Improve styling and layout in AuthDialog component
This commit refines the styling in the AuthDialog component by adjusting the border-radius of the otpInput elements in the associated Sass module. Additionally, it enhances the appearance of the Snackbar component by modifying the margin and padding of the Alert components within it. These changes improve the overall aesthetics and user experience of the AuthDialog.