Add Footer.interface.ts defining FooterProps interface
This commit introduces a new file Footer.interface.ts to the project. The file defines the FooterProps interface, which specifies the structure of props expected by the Footer component. The FooterProps interface includes a property links of type NavigationLinks, which is imported from the Navigation.type.ts file. This interface will be used to pass navigation links to the Footer component.