Add Header.interface.ts defining HeaderProps interface
This commit introduces a new file Header.interface.ts to the project. The file defines the HeaderProps interface, which specifies the structure of props expected by the Header component. The HeaderProps interface includes a property links of type NavigationLinks, which is imported from the Navigation.type.ts file. This interface will be used to pass navigation links to the Header component.