Refactor Footer component to accept navigation links as props
This commit refactors the Footer component to accept navigation links as props, following the defined FooterProps interface. The links are mapped over in the component to dynamically generate the list items for quick access to different pages. This change improves the flexibility and reusability of the Footer component by allowing it to display any set of navigation links passed to it.